+++ lib/src/bones_api_entity_db_object_memory.dart
@@ -285,7 +285,8 @@ class DBObjectMemoryAdapter
   }
 
   @override
-  FutureOr<bool> isConnectionValid(connection) => true;
+  FutureOr<bool> isConnectionValid(connection, {bool checkUsage = true}) =>
+      true;
 
   final Map<DBObjectMemoryAdapterContext, DateTime> _openTransactionsContexts =
       <DBObjectMemoryAdapterContext, DateTime>{};